requirements.txt
fastapi
uvicorn
httpx
cachetools
pydantic
tenacity
python-dotenv